aiopillow/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiocache==0.11.1
    # via -r requirements.in
aioconsole==0.4.1
    # via aiomonitor
aiodebug==2.3.0
    # via -r requirements.in
aiofile==3.7.4
    # via aiopath
aiofiles==0.8.0
    # via -r requirements.in
aiohttp==3.8.1
    # via
    #   -r requirements.in
    #   aiohttp-json-rpc
aiohttp-json-rpc==0.13.3
    # via -r requirements.in
aiomonitor==0.4.5
    # via -r requirements.in
aiopath==0.5.12
    # via -r requirements.in
aioreactive==0.15.0
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
aiosql==3.4.1
    # via -r requirements.in
aiosqlite==0.17.0
    # via -r requirements.in
anyio==3.5.0
    # via
    #   aiopath
    #   asyncer
    #   starlette
appnope==0.1.2
    # via ipython
asttokens==2.0.5
    # via snoop
async-timeout==4.0.2
    # via aiohttp
asyncer==0.0.1
    # via -r requirements.in
attrs==21.4.0
    # via
    #   -r requirements.in
    #   aiohttp
backcall==0.2.0
    # via ipython
backoff==1.11.1
    # via -r requirements.in
bcrypt==3.2.0
    # via passlib
beautifulsoup4==4.10.0
    # via
    #   -r requirements.in
    #   bs4
    #   google
better-exceptions==0.3.3
    # via -r requirements.in
bs4==0.0.1
    # via pyinspect
caio==0.9.5
    # via aiofile
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via bcrypt
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
cheap-repr==0.5.1
    # via snoop
click==7.1.2
    # via
    #   pip-tools
    #   pyinspect
    #   scenedetect
    #   typer
click-spinner==0.1.10
    # via -r requirements.in
cloudpickle==2.0.0
    # via dask
codetiming==1.3.0
    # via -r requirements.in
colorama==0.4.4
    # via
    #   pretty-errors
    #   rich
    #   typer-cli
commonmark==0.9.1
    # via rich
contextlib2==21.6.0
    # via aiosql
cycler==0.11.0
    # via matplotlib
dask==2022.2.0
    # via -r requirements.in
decorator==5.1.1
    # via
    #   ipython
    #   validators
dnspython==2.2.0
    # via email-validator
docutils==0.16
    # via -r requirements.in
durabledict==0.9.4
    # via gutter
email-validator==1.1.3
    # via
    #   -r requirements.in
    #   pydantic
executing==0.8.2
    # via snoop
expression==1.2.0
    # via aioreactive
factory-boy==3.2.1
    # via -r requirements.in
faker==13.0.0
    # via factory-boy
fastapi==0.74.0
    # via prometheus-fastapi-instrumentator
fonttools==4.29.1
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.1.0
    # via dask
gallery-dl==1.20.5
    # via -r requirements.in
google==3.0.0
    # via pyinspect
gutter==0.5.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
    #   yarl
imageio==2.16.0
    # via -r requirements.in
importlib-metadata==4.11.1
    # via markdown
imutils==0.5.4
    # via -r requirements.in
ipython==7.1
    # via -r requirements.in
jedi==0.18.1
    # via ipython
jsonpickle==2.1.0
    # via gutter
kiwisolver==1.3.2
    # via matplotlib
locket==0.2.1
    # via partd
logging-tree==1.9
    # via -r requirements.in
loguru==0.6.0
    # via -r requirements.in
markdown==3.3.6
    # via -r requirements.in
matplotlib==3.5.1
    # via -r requirements.in
memory-profiler==0.60.0
    # via -r requirements.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.22.2
    # via
    #   imageio
    #   matplotlib
    #   opencv-python
    #   pyinspect
    #   scenedetect
opencv-python==4.5.5.62
    # via scenedetect
packaging==21.3
    # via
    #   dask
    #   matplotlib
    #   pytesseract
parso==0.8.3
    # via jedi
partd==1.2.0
    # via dask
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.0.1
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   pytesseract
pip-tools==6.5.1
    # via -r requirements.in
pretty-errors==1.2.25
    # via -r requirements.in
prometheus-client==0.13.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==5.7.1
    # via -r requirements.in
prompt-toolkit==2.0.10
    # via ipython
psutil==5.9.0
    # via memory-profiler
ptyprocess==0.7.0
    # via pexpect
pycparser==2.21
    # via cffi
pydantic[email]==1.9.0
    # via
    #   -r requirements.in
    #   fastapi
pygments==2.7.4
    # via
    #   -r requirements.in
    #   ipython
    #   rich
    #   snoop
pyinspect==0.1.0
    # via -r requirements.in
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pytesseract==0.3.9
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   faker
    #   matplotlib
python-dotenv==0.19.2
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
python-slugify==6.0.1
    # via -r requirements.in
pytz==2021.3
    # via -r requirements.in
pyyaml==5.4.1
    # via dask
requests==2.27.1
    # via
    #   -r requirements.in
    #   gallery-dl
    #   pyinspect
rich==11.2.0
    # via
    #   -r requirements.in
    #   pyinspect
scenedetect[opencv]==0.5.6.1
    # via -r requirements.in
shellingham==1.4.0
    # via typer-cli
six==1.16.0
    # via
    #   asttokens
    #   bcrypt
    #   prompt-toolkit
    #   python-dateutil
    #   python-multipart
    #   snoop
    #   validators
sniffio==1.2.0
    # via anyio
snoop==0.4.1
    # via -r requirements.in
soupsieve==2.3.1
    # via beautifulsoup4
starlette==0.17.1
    # via fastapi
telnetlib3==1.0.4
    # via -r requirements.in
tenacity==8.0.1
    # via -r requirements.in
terminaltables==3.1.10
    # via aiomonitor
text-unidecode==1.3
    # via python-slugify
tomli==1.2.3
    # via
    #   -r requirements.in
    #   pep517
toolz==0.11.2
    # via
    #   dask
    #   partd
tqdm==4.62.3
    # via
    #   -r requirements.in
    #   scenedetect
traitlets==5.1.1
    # via ipython
typer==0.3.2
    # via
    #   -r requirements.in
    #   typer-cli
typer-cli==0.0.12
    # via -r requirements.in
typing-extensions==4.1.1
    # via
    #   -r requirements.in
    #   aiodebug
    #   aiopath
    #   aiosql
    #   aiosqlite
    #   pydantic
unidecode==1.3.2
    # via -r requirements.in
uritemplate==4.1.1
    # via -r requirements.in
uritools==4.0.0
    # via -r requirements.in
urllib3==1.26.8
    # via requests
validators==0.18.2
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.3
    # via gutter
wheel==0.37.1
    # via pip-tools
wrapt==1.13.3
    # via -r requirements.in
yarl==1.7.2
    # via aiohttp
youtube-dl==2021.12.17
    # via -r requirements.in
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
